\name{tempEng}
\alias{tempEng}
\docType{data}
\title{Time series of monthly temperature values}
\description{
  Monthly temperature in Central England from 1723-1970
}
\usage{data(tempEng)}
\format{
 Time-Series [1:2976] from 1723 to 1971
}

\source{Hipel, K. W. and Mcleod, A. (1994) Time Series Modelling of Water Resources and Environmental Systems, Elsevier  \cr\cr
}

\examples{
data(tempEng)
%## maybe str(tempEng) ; plot(tempEng) ...
}
\keyword{datasets}
